refactor PatchCore implementation
