+++ anomalib/models/patchcore/__init__.py
@@ -13,3 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions
 # and limitations under the License.
+
+from .model import PatchcoreLightning
+
+__all__ = ["PatchcoreLightning"]
